requirements.txt
giskard>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
pathlib2>=2.3.0
google-generativeai>=0.3.0
